refactor get transaction info method
