+++ app/src/main/java/com/xpay/kotlinutils/model/TransactionData.kt
@@ -1,4 +1,4 @@
-package com.xpay.kotlin.models
+package com.xpay.kotlinutils.model
 
 import com.google.gson.annotations.SerializedName
 
@@ -12,7 +12,6 @@ data class TransactionData (
     @SerializedName("payment_for") val payment_for : String,
     @SerializedName("quantity") val quantity : String,
     @SerializedName("status") val status : String,
-    @SerializedName("community") val community : Community,
     @SerializedName("custom_fields_json") val custom_fields_json : String,
     @SerializedName("total_amount_piasters") val total_amount_piasters : Int
 )
